% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ancom_functions.r
\name{ancom.detect}
\alias{ancom.detect}
\title{Calculate test statistic for ANCOM}
\usage{
ancom.detect(otu_data, n_otu, alpha, multcorr, ncore)
}
\arguments{
\item{otu_data}{the OTU dataset.}

\item{n_otu}{the number of OTUs.}

\item{alpha}{the significance level at which the tests are to be performed.}

\item{multcorr}{type of correction for multiple comparisons, see Details.}

\item{ncore}{if ncore>1, then \pkg{doParallel} will be loaded and used.}

\item{Wexact}{logical, should Wilcoxon tests return exact p-values?}
}
\description{
Calculates test statistics for differences in OTU abundances between treatment groups.
}
\details{
\code{multcorr} can take on values of 1 (no correction), 2 (a less stringent)
correction, or 3 (a more stringent correction).
}
\note{
This function is intended to be called by \code{\link{ANCOM}}, see the documentation of
that function for details on using the method.
}

