% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packageAndData.r
\docType{package}
\name{ancom.R-package}
\alias{ancom.R-package}
\title{ancom.R}
\description{
The ANCOM method.
}
\details{
\tabular{ll}{
Package: \tab ancom.R\cr
Type:    \tab Package\cr
Version: \tab 1.1-2\cr
Date:    \tab 2015-07-09\cr
License: \tab GLP-3  \cr
}
}

