% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ancom.r
\name{plot_ancom}
\alias{plot_ancom}
\title{Plot of data from objects of class 'ancom'}
\usage{
plot_ancom(object, ncols = -1, ...)
}
\arguments{
\item{object}{object of class \code{ancom}.}

\item{ncols}{the number of columns for \code{ggplot} to produce using \code{facet_wrap}.
If \code{ncol=-1}, then the function will attempt to}

\item{...}{space for additional arguments (none corrently)}
}
\description{
Produces comparison boxplots of data for objects of class \code{ancom}.
}
\details{
\code{plot_ancom} uses \pkg{ggplot} to produce graphics.
}

