% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_ancom.r, R/shiny_ancomServer.r, R/shiny_ancomUI.r
\docType{data}
\name{shiny_ancom}
\alias{shiny_ancom}
\alias{shiny_ancomServer}
\alias{shiny_ancomUI}
\title{Shiny GUI for ancom.R}
\format{\preformatted{List of 3
 $ : NULL
 $ : NULL
 $ :List of 1
  ..$ :List of 3
  .. ..$ name    : chr "div"
  .. ..$ attribs :List of 1
  .. .. ..$ class: chr "container-fluid"
  .. ..$ children:List of 2
  .. .. ..$ :List of 2
  .. .. .. ..$ :List of 3
  .. .. .. .. ..$ name    : chr "head"
  .. .. .. .. ..$ attribs : list()
  .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. ..$ name    : chr "title"
  .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. ..$ : chr "Analysis of Composition of Microbiomes (ANCOM) v1.1-3"
  .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..$ :List of 3
  .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. ..$ class: chr "col-sm-12"
  .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. ..$ name    : chr "h1"
  .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. ..$ : chr "Analysis of Composition of Microbiomes (ANCOM) v1.1-3"
  .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
  .. .. ..$ :List of 3
  .. .. .. ..$ name    : chr "div"
  .. .. .. ..$ attribs :List of 1
  .. .. .. .. ..$ class: chr "row"
  .. .. .. ..$ children:List of 2
  .. .. .. .. ..$ :List of 3
  .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. ..$ class: chr "col-sm-4"
  .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. ..$ name    : chr "form"
  .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. ..$ class: chr "well"
  .. .. .. .. .. .. .. ..$ children:List of 15
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Choose data file"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 4
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id    : chr "file1"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ name  : chr "file1"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type  : chr "file"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ accept: chr "text/csv,text/comma-separated-values,text/plain,.csv"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "file1_progress"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "progress progress-striped active shiny-file-input-progress"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "progress-bar"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "Accepted file formats: csv, txt"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "file_out"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Path of output file (without file name)"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 4
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "file_out"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "text"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-control"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr ""
  .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. ..$ style: chr "width: 90\%;"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "datafmt"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Format of input dataset:"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "datafmt"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :Classes 'html', 'character'  atomic [1:1] <option value="wide" selected>Subjects on rows, OTUs on columns</option>
<option value="tall">OTUs on rows, Subjects on columns</option>
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "datafmt"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "\{\}"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 8
  .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.11.2"
  .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
  .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap3.css"
  .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\\n<script src=\\"shared/selectize/js/es5-shim.min.js\\"></script>\\n<![endif]-->\\n<script src=\\"shared/selectize/"| __truncated__
  .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "repeated"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Repeated measures"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "adjust"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Correct for multiple testing"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.adjust == true "
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "fdr"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Enter desired FDR (between 0 and 1):"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 7
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "fdr"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-control"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "0.05"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ step : num 0.005
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.adjust == false "
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "alpha"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Enter significance level (between 0 and 1):"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 7
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "alpha"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-control"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "0.05"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 0
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ step : num 0.005
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id  : chr "fixPlot"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type: chr "checkbox"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Adjust figure:"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ data-display-if: chr "input.fixPlot == true "
  .. .. .. .. .. .. .. .. .. ..$ children:List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "pltWidth"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Width of plot (\%)"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 7
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "pltWidth"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-control"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "100"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 10
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 100
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ step : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for: chr "pltHeight"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Height of plot (pixel)"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 7
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "pltHeight"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type : chr "number"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-control"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ value: chr "400"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ min  : num 100
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ max  : num 10000
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ step : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
  .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "ncols"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Number of columns for the plot"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 12
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class                  : chr "js-range-slider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id                     : chr "ncols"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-min               : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-max               : num 10
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-from              : num 3
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-step              : num 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid              : chr "true"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-num          : num 9
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-snap         : chr "false"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-prettify-separator: chr ","
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-keyboard          : chr "true"
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-keyboard-step     : num 11.1
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 8
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "ionrangeslider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "2.0.6"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
  .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/ionrangeslider"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "js/ion.rangeSlider.min.js"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr [1:2] "css/ion.rangeSlider.css" "css/ion.rangeSlider.skinShiny.css"
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
  .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "hr"
  .. .. .. .. .. .. .. .. .. ..$ attribs : list()
  .. .. .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "compute1"
  .. .. .. .. .. .. .. .. .. .. ..$ type : chr "button"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "btn btn-default action-button"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Run ANCOM"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
  .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "compute2"
  .. .. .. .. .. .. .. .. .. .. ..$ type : chr "button"
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "btn btn-default action-button"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ :List of 2
  .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Update Plot"
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. .. .. ..$ name    : chr "span"
  .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ class: chr "help-block"
  .. .. .. .. .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. .. .. .. .. ..$ : chr "'Update Plot' must be clicked before the plot will appear."
  .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. ..$ :List of 3
  .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. ..$ attribs :List of 1
  .. .. .. .. .. .. ..$ class: chr "col-sm-8"
  .. .. .. .. .. ..$ children:List of 1
  .. .. .. .. .. .. ..$ :List of 3
  .. .. .. .. .. .. .. ..$ name    : chr "div"
  .. .. .. .. .. .. .. ..$ attribs :List of 2
  .. .. .. .. .. .. .. .. ..$ id   : chr "theTabset"
  .. .. .. .. .. .. .. .. ..$ class: chr "shiny-html-output"
  .. .. .. .. .. .. .. ..$ children: list()
  .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. .. .. ..- attr(*, "class")= chr "shiny.tag"
  .. ..- attr(*, "class")= chr "shiny.tag"
 - attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
 - attr(*, "html_dependencies")=List of 1
  ..$ :List of 8
  .. ..$ name      : chr "bootstrap"
  .. ..$ version   : chr "3.3.1"
  .. ..$ src       :List of 2
  .. .. ..$ href: chr "shared/bootstrap"
  .. .. ..$ file: chr "/home/statman/R/x86_64-pc-linux-gnu-library/3.2/shiny/www/shared/bootstrap"
  .. ..$ meta      :List of 1
  .. .. ..$ viewport: chr "width=device-width, initial-scale=1"
  .. ..$ script    : chr [1:3] "js/bootstrap.min.js" "shim/html5shiv.min.js" "shim/respond.min.js"
  .. ..$ stylesheet: chr "css/bootstrap.min.css"
  .. ..$ head      : NULL
  .. ..$ attachment: NULL
  .. ..- attr(*, "class")= chr "html_dependency"
}}
\usage{
shiny_ancom()

shiny_ancomServer(input, output)

shiny_ancomUI
}
\arguments{
\item{input}{input from GUI.}

\item{output}{output to GUI.}
}
\description{
Opens a graphical user interface to run \pkg{ancom.R}, built from the \pkg{shiny} package.




}
\details{
This GUI is the primary purpose for the package \pkg{ancom.R}.
}
\examples{
\dontrun{ shiny_ancom() }
}
\keyword{datasets}

